[류] refactor: correction_module을 moduls 폴더로 이동 | 모듈 파일 위치 정리, import 경로 수정 | 2026-04-16
